+++ isvg_im_lib/entities/im_object.psm1
@@ -1,0 +1,5 @@
+class IM_Object{
+
+	hidden IM_Object () {}
+
+}