+++ isvg_im_lib/utils/utils_proxy.psm1
@@ -1,3 +1,5 @@
+using module "..\utils\utils_logs.psm1"
+
 function Copy-ISIMObjectNamespace {
     <#
     .SYNOPSIS
@@ -25,7 +27,7 @@ function Copy-ISIMObjectNamespace {
     if( $obj.getType().BaseType.Name -eq "Array" ) {
         $tmp_array	=	@();
         
-        $obj | % {
+        $obj | ForEach-Object {
             $tmp1	=	Copy-ISIMObjectNamespace $_ $targetNS;
             $tmp_array += $tmp1;
         }
@@ -36,10 +38,10 @@ function Copy-ISIMObjectNamespace {
 
     $newObj	=	New-Object ( $targetNS+"."+$myTypeName)
 
-    $obj.psobject.Properties | % {
+    $obj.psobject.Properties | ForEach-Object {
         $pname	=	$_.Name
         if ( $_.TypeNameOfValue.StartsWith("System.") ) {
-            if( $newObj.psobject.Properties.Item($pname) -ne $null ) {
+            if( $null -ne $newObj.psobject.Properties.Item($pname) ) {
                 $newObj.$pname	=	$_.Value
             } else {
                 Write-Host -ForegroundColor Yellow "Property $pname Could not be set"
@@ -48,10 +50,17 @@ function Copy-ISIMObjectNamespace {
             if ( !$newObj.$pname ) {
                 $newObj.$pname	=	New-Object ( $targetNS+"."+($_.TypeNameOfValue.Split(".")[-1].Split("[")[0]))
             }
-            $newObj.$pname	=	Copy-ISIMObjectNamespace $obj.$pname $targetNS
-        }
+            # ISSUE #2:    $obj is null when converting container raw from role_proxy
+            # Detail:   Once pname = "children" cannot keep looping correctly. Somewhere $obj is null
+            # TODO:     Keep an eye on this funtion to avoid new errors due to this IF
+            if ($pname -ne "children"){
+                $newObj.$pname	=	Copy-ISIMObjectNamespace $obj.$pname $targetNS
+            }else{
+                Write-Warning "Children property skipped"
+            }
     }
-    return $newObj
+}
+return $newObj
 }
 
 function Convert-2WSAttr {
@@ -72,10 +81,10 @@ function Convert-2WSAttr {
         $inAttr
     )
     process {
-        if ( $inAttr -NE $null ) {
+        if ( $null -NE $inAttr ) {
             $wsattr_array	=	$inAttr;
 
-            $hash.GetEnumerator() | ForEach{
+            $hash.GetEnumerator() | ForEach-Object{
                 if ( $null -ne $_.value ){
                     $prop_name	=	$_.name;
                     $prop_value	=	$_.value;
@@ -93,7 +102,7 @@ function Convert-2WSAttr {
 
         } else {
             $wsattr_array	=	@();
-            $hash.GetEnumerator() | ForEach{
+            $hash.GetEnumerator() | ForEach-Object{
                 if ( $null -ne $_.value ){
                     $wsattr	=	New-Object ($namespace+".WSAttribute")
                     $wsattr.name	=	$_.name
@@ -107,7 +116,7 @@ function Convert-2WSAttr {
 
 }
 
-function Convert-2WSObject{
+function Convert-2WSObject {
     <#
     .SYNOPSIS
         Helper Function to manage WSObject with Custom ISIM objects
@@ -123,6 +132,10 @@ function Convert-2WSObject{
         [psObject]$wsStub
     )
 
+    # Remove empty attributes
+    @($isim_object.attributes.keys) | ForEach-Object { if (-not $isim_object.attributes[$_]) { $isim_object.attributes.Remove($_) } }
+    
+    # Clone isim_object to a WSObject
     $isim_object_props         	=	  $isim_object | Get-Member | Where-Object{ ( $_.membertype -eq "Property" )}
 
     $isim_object_props | ForEach-Object {
@@ -135,5 +148,65 @@ function Convert-2WSObject{
             $wsStub.attributes 	=	  Convert-2WSAttr -hash $isim_object.attributes -namespace $wsStub.GetType().Namespace
         }
     }
+    return $wsStub
+}
 
+function Convert-Hash2WSAttr {
+    <#
+    .SYNOPSIS
+        Helper Function to manage WSAttr with Hash Tables
+    .DESCRIPTION
+        Helper Function to manage WSAttr with Hash Tables. Will generate an WSAttr Object by adding Values from a Hash Table.
+    #>
+    [CmdletBinding()]
+    [OutputType([string])]
+    param (
+        [Parameter(Mandatory=$true)]
+        [hashtable]$hash,
+        [Parameter(Mandatory=$true)]
+        [string]$namespace,
+        [Parameter(ValueFromPipelineByPropertyName=$true)]
+        $inAttr
+    )
+    process {
+
+
+        if ( $null -NE $inAttr ) {
+            $wsattr_array = $inAttr;
+
+            $hash.GetEnumerator() | ForEach-Object{
+                $prop_name = $_.name;
+                $prop_value = $_.value;
+
+	            if ( ( $wsattr_array | Where-Object { $_.name -eq $prop_name }).Count -eq 1 ) {
+		            ( $wsattr_array | Where-Object { $_.name -eq $prop_name }).values = $prop_value
+	            } else {
+                    $wsattr = New-Object ($namespace+".WSAttribute")
+                    $wsattr.name = $prop_name
+                    $wsattr.values +=  $prop_value
+                    $wsattr_array += $wsattr
+                }
+
+            }
+
+        } else {
+            $wsattr_array = @();
+            $hash.GetEnumerator() | ForEach-Object{
+                $wsattr = New-Object ($namespace+".WSAttribute")
+                $wsattr.name = $_.name
+                $wsattr.values +=  $_.value
+                $wsattr_array += $wsattr
+            }
+        }
+
+
+
+        return $wsattr_array;
+    }
+}
+
+
+
+function to_WSObject() {
+		
 }